# lib/swagger/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shop API",
    "description": "مستندات کامل و دقیق API فروشگاه بر پایه Express و MongoDB",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/brand": {
      "get": {
        "tags": ["Brands"],
        "summary": "دریافت لیست تمام برندها",
        "responses": {
          "200": {
            "description": "لیست برندها",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Brand" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Brands"],
        "summary": "ایجاد برند جدید",
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddBrandRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "برند با موفقیت ایجاد شد",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "برند با موفقیت ایجاد شد" },
                    "data": { "$ref": "#/components/schemas/Brand" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "عدم دسترسی احراز هویت",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "برندی با این نام قبلاً ثبت شده است",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "اطلاعات ارسالی نامعتبر است",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/category": {
      "get": {
        "tags": ["Categories"],
        "summary": "دریافت لیست تمام دسته‌بندی‌ها",
        "responses": {
          "200": {
            "description": "لیست دسته‌بندی‌ها",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Category" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "ایجاد دسته‌بندی جدید",
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddCategoryRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "دسته‌بندی با موفقیت ایجاد شد",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "دسته‌بندی با موفقیت ایجاد شد" },
                    "data": { "$ref": "#/components/schemas/Category" }
                  }
                }
              }
            }
          },
          "401": { "description": "عدم دسترسی", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "دسته‌بندی با این نام یا نام انگلیسی قبلاً ثبت شده است", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "اطلاعات ارسالی نامعتبر است", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } }
        }
      }
    },
...
    "/api/media": {
      "get": {
        "tags": ["Media"],
        "summary": "دریافت لیست تمام فایل‌ها و رسانه‌ها",
        "responses": {
          "200": {
            "description": "لیست رسانه‌ها",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Media" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Media"],
        "summary": "ایجاد رسانه جدید",
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddMediaRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "رسانه با موفقیت ایجاد شد",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "رسانه با موفقیت ایجاد شد" },
                    "data": { "$ref": "#/components/schemas/Media" }
                  }
                }
              }
            }
          },
          "401": { "description": "عدم دسترسی", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "آدرس رسانه نامعتبر است", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "هدر استاندارد Authorization به صورت Bearer <token>"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "کوکی نشست با کلید token"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "منبع مورد نظر یافت نشد" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "اطلاعات ارسالی نامعتبر است" },
          "errors": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["نام باید حداقل ۳ کاراکتر باشد", "لطفاً یک آدرس ایمیل معتبر وارد کنید"]
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            },
            "example": {
              "name": ["نام باید حداقل ۳ کاراکتر باشد"],
              "email": ["لطفاً یک آدرس ایمیل معتبر وارد کنید"]
            }
          }
        }
      },
...
      "Brand": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65f1a2b3c4d5e6f7a8b9c0d2" },
          "name": { "type": "string", "example": "سامسونگ" },
          "logo": { "type": "string", "format": "uri", "example": "https://example.com/samsung.png" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AddBrandRequest": {
        "type": "object",
        "required": ["name", "logo"],
        "properties": {
          "name": { "type": "string", "example": "سامسونگ" },
          "logo": { "type": "string", "format": "uri", "example": "https://example.com/samsung.png" }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65f1a2b3c4d5e6f7a8b9c0d3" },
          "name": { "type": "string", "example": "گوشی موبایل" },
          "en_name": { "type": "string", "example": "mobile" },
          "image": { "type": "string", "format": "uri", "example": "https://example.com/mobile.jpg" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AddCategoryRequest": {
        "type": "object",
        "required": ["name", "en_name", "image"],
        "properties": {
          "name": { "type": "string", "example": "گوشی موبایل" },
          "en_name": { "type": "string", "example": "mobile" },
          "image": { "type": "string", "format": "uri", "example": "https://example.com/mobile.jpg" }
        }
      },
...
      "Media": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65f1a2b3c4d5e6f7a8b9c0d4" },
          "url": { "type": "string", "format": "uri", "example": "https://example.com/images/banner.jpg" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AddMediaRequest": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": { "type": "string", "format": "uri", "example": "https://example.com/images/banner.jpg" }
        }
      },
...
    }
  }
}